Fixed a couple more warnings on Linux x86_64
git-svn-id: https://hypertable.googlecode.com/svn/trunk@469 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Hypertable/RangeServer/CommitLogReader.cc
@@ -93,7 +93,7 @@ CommitLogReader::CommitLogReader(Filesystem *fs, std::string &logDir) : m_fs(fs)
       exit(1);
     }
 
-    if (flen < sizeof(CommitLogHeaderT))
+    if (flen < (int64_t)sizeof(CommitLogHeaderT))
       continue;
 
     if ((error = m_fs->open(m_log_file_info[i].fname, &fd)) != Error::OK) {
